Update documentation with controls
